+++ src/photon_platform/clerk/curator/__init__.py
@@ -1,0 +1,16 @@
+"""
+curator
+=======
+
+PHOTON platform
+---------------
+
+
+"""
+from photon_platform.curator.curator import Curator
+
+__author__ = "PHOTON platform"
+__maintainer__ = "PHOTON platform"
+__email__ = "github@phiarchitect.com"
+__version__ = '0.0.3'
+__licence__ = "MIT"