+++ src/photon_platform/clerk/gather/app.py
@@ -1,0 +1,9 @@
+"""
+run the main app
+"""
+from .gather import Gather
+
+
+def run() -> None:
+    reply = Gather().run()
+    print(reply)