+++ src/photon_platform/clerk/logger/__init__.py
@@ -1,0 +1,18 @@
+"""
+logger
+======
+
+PHOTON platform
+---------------
+
+tools for creating log entries for Sphinx-based content sites.
+
+"""
+
+from .logger import Logger
+
+__author__ = "phiarchitect"
+__maintainer__ = "PHOTON platform"
+__email__ = "github@phiarchitect.com"
+__version__ = "0.0.1"
+__licence__ = "MIT"