+++ src/photon_platform/clerk/progenitor/__init__.py
@@ -1,0 +1,18 @@
+"""
+progenitor
+==========
+
+PHOTON platform
+---------------
+
+tools for creating and initializing Python projects
+
+"""
+
+from photon_platform.progenitor.progenitor import create_project
+
+__author__ = "phiarchitect"
+__maintainer__ = "PHOTON platform"
+__email__ = "github@phiarchitect.com"
+__version__ = "0.0.1"
+__licence__ = "MIT"